Move function to App.js
